extract editor setup logic into «useSetup» hook
Improve code organization and reusability by moving editor initialization logic from the main component into a dedicated «useSetup» hook.
Simplifies the main component and promotes separation of concerns.
